refactor: extract common Popover logic in a custom hook
